elf2tbf: rewrite
- Refactor elf2tbf to be a little cleaner. There is probably a lot to
improve still but it is a little bit more structured.
- Move the crt0 header stuff to elf2tbf. This allows us to use that same
method with other runtimes, but also means there is less tension between
the linker script and elf2tbf. Before we were doing a lot in the linker
file, and then just trusting elf2tbf to not override it, yet it is
entirely in elf2tbf's purview to do so. This moves basically everything
to elf2tbf so we can centralize all of the logic.